Renamed old BeforeAndAfterSuite to BeforeAndAfterEachAllSuite, and BeforeNAfterSuite to BeforeAndAfterSuite.
